% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad.R
\name{lorad}
\alias{lorad}
\title{This function calculates the Lorad estimate of the marginal likelihood}
\usage{
lorad(params, colspec, training_frac, training_mode, coverage, file_name)
}
\value{
Lorad estimate of marginal likelihood
}
\description{
This function calculates the Lorad estimate of the marginal likelihood
}
