% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-calc-log-sum.R
\name{lorad_calc_log_sum}
\alias{lorad_calc_log_sum}
\title{Calculate a sum on log scale}
\usage{
lorad_calc_log_sum(logx)
}
\arguments{
\item{logx}{Numeric vector in which elements are on log scale}
}
\value{
The log of the sum of the (exponentiated) elements supplied in logx
}
\description{
Calculates the (natural) log of a sum without leaving the log scale by factoring
out the largest element.
}
\examples{
# log of sum 1 + 2 + 3 is log(6) = 1.79175947
v <- c(log(1), log(2), log(3))
lorad_calc_log_sum(v)
}
