% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-estimate.R
\name{lorad_estimate}
\alias{lorad_estimate}
\title{Calculates the LoRaD estimate of the marginal likelihood}
\usage{
lorad_estimate(params, colspec, training_frac, training_mode, coverage)
}
\arguments{
\item{params}{Data frame in which rows are sample points and columns are parameters, except that last column holds the log posterior kernel}

\item{colspec}{Named character vector associating column names in params with column specifications}

\item{training_frac}{Number between 0 and 1 specifying the training fraction}

\item{training_mode}{One of random, left, or right, specifying how training fraction is chosen}

\item{coverage}{Number between 0 and 1 specifying fraction of training sample used to compute working parameter space}
}
\value{
LoRaD estimate of log marginal likelihood
}
\description{
Calculates the LoRaD estimate of the marginal likelihood
}
