% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-standardize.R
\name{lorad_standardize}
\alias{lorad_standardize}
\title{Standardizes parameters that have already been transformed (if necessary) so that they
are unconstrained. Standardization involves subtracting the mean for a particular parameter
from each of its values and dividing by the standard deviation. Assumes that the log kernel
is the last column in the supplied data frame.}
\usage{
lorad_standardize(df, coverage)
}
\arguments{
\item{df}{Data frame containing a column for each model parameter sampled and a final column of log posterior kernel values}

\item{coverage}{Fraction of the training sample used to compute working parameter space}
}
\value{
List containing the log-Jacobian of the standardization transformation, the inverse square root matrix, a vector of column means, and rmax (radial distance to furthest point in working parameter space)
}
\description{
Standardizes parameters that have already been transformed (if necessary) so that they
are unconstrained. Standardization involves subtracting the mean for a particular parameter
from each of its values and dividing by the standard deviation. Assumes that the log kernel
is the last column in the supplied data frame.
}
