% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-standardize-estimation-sample.R
\name{lorad_standardize_estimation_sample}
\alias{lorad_standardize_estimation_sample}
\title{Transforms training sample using training sample means and standard deviations}
\usage{
lorad_standardize_estimation_sample(standardinfo, y)
}
\arguments{
\item{standardinfo}{List containing the log Jacobian of the standardization
transformation, the inverse square root matrix, the column means, and
rmax (the radial distance representing the edge of the working parameter space)}

\item{y}{Data frame containing a column for each transformed model parameter
in the estimation sample, with last column being the log kernel values}
}
\value{
A new data frame consisting of the standardized estimation sample with
log kernel in last column
}
\description{
Transforms training sample using training sample means and standard deviations
}
