% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-transform.R
\name{lorad_transform}
\alias{lorad_transform}
\title{Takes a data frame consisting of sampled parameter values
and a named character vector associating a column name
in the data frame to a column specification and
returns a new data frame containing parameters that
have been log-transformed if originally constrained.}
\usage{
lorad_transform(params, colspec)
}
\arguments{
\item{params}{Data frame containing a column for each model parameter sampled as well as columns that constitute the log posterior kernel}

\item{colspec}{Named character vector associating each column name in params to a column specification}
}
\value{
A new data frame consisting of transformed parameter values
}
\description{
Takes a data frame consisting of sampled parameter values
and a named character vector associating a column name
in the data frame to a column specification and
returns a new data frame containing parameters that
have been log-transformed if originally constrained.
}
