% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbit-function.R
\name{rabbit_function}
\alias{rabbit_function}
\title{A Rabbit Function}
\usage{
rabbit_function(love = TRUE)
}
\arguments{
\item{love}{Do you love rabbits? Defaults to TRUE.}
}
\description{
This function allows you to express your love of rabbits.
}
\examples{
rabbit_function()
}
\keyword{rabbits}
