% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{This function takes a data frame y, that contains the log transform parameter values
The log kernel is always the last column}
\usage{
standardize(df, coverage)
}
\value{
A list of standardize info of logJ, invsqrts, colMeans(x), rmax
}
\description{
This function takes a data frame y, that contains the log transform parameter values
The log kernel is always the last column
}
