% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-estimation-sample.R
\name{standardize_estimation_sample}
\alias{standardize_estimation_sample}
\title{This function standardized the estimation sample}
\usage{
standardize_estimation_sample(standardinfo, y)
}
\arguments{
\item{standardinfo}{Contains list of logJ, invsqrts, colMeans(x), rmax}

\item{y}{data frame containing a column for each transformed model parameter sampled}
}
\value{
A new data frame consisting of standardized estimation sample with log kernal in last column
}
\description{
\code {standardize_estimation_sample} A function that tajes the parameters standardinfo and y
}
