% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLogSum.R
\name{calcLogSum}
\alias{calcLogSum}
\title{This function calculates log of a sum without leaving the log scale}
\usage{
calcLogSum(logx)
}
\description{
This function calculates log of a sum without leaving the log scale
}
