% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-data.R
\docType{data}
\name{k80samples}
\alias{k80samples}
\title{Sequence data used in k80 vignette}
\format{
\subsection{\code{k80samples}}{

A data frame with 10,000 rows and 4 columns:
\describe{
\item{iter}{Iteration}
\item{log.kernel}{Log unnormalized posterior}
\item{edgelen}{Edge length in substitutions per site}
\item{kappa}{Transition transversion rate ratio}
}
}
}
\source{
\doi{10.1093/sysbio/syad007}
}
\usage{
k80samples
}
\description{
Sequence data used in k80 vignette
}
\keyword{datasets}
