% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad.R
\name{lorad}
\alias{lorad}
\title{This function calculates the Lorad estimate of the marginal likelihood}
\usage{
lorad(params, colspec, training_frac, training_mode, coverage, file_name)
}
\arguments{
\item{params}{Parameters, log kernel is always the last column}

\item{colspec}{Identification of columns}

\item{training_frac}{Specified training fraction}

\item{training_mode}{Training mode is random, left, or right}

\item{coverage}{Fraction of training sample used to compute working parameter space}

\item{file_name}{Name of parameter file}
}
\value{
LoRaD estimate of marginal likelihood
}
\description{
This function calculates the Lorad estimate of the marginal likelihood
}
