% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorad-calc-log-sum.R
\name{lorad_calc_log_sum}
\alias{lorad_calc_log_sum}
\title{Calculate a sum on log scale}
\usage{
lorad_calc_log_sum(logx)
}
\arguments{
\item{logx}{Numeric vector in which elements are on log scale}
}
\value{
The log of the sum of the (exponentiated) elements supplied in logx
}
\description{
Calculates the (natural) log of a sum without leaving the log scale by factoring
out the largest element.
}
