% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{This function takes a data frame y, that contains the log transform parameter values
The log kernel is always the last column}
\usage{
standardize(df, coverage)
}
\arguments{
\item{df}{data frame containing a column for each model parameter sampled as well as columns that constitute the log posterior kernel}

\item{coverage}{Fraction of training sample used to compute working parameter space}
}
\value{
A list of standardize info of logJ, invsqrts, colMeans(x), rmax
}
\description{
\code {standardize} Function in which the parameters are df, a data frame, and coverage
}
