% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{This function takes a params data frame consisting of sampled parameter values
and a colspec data frame that specifies what each column in the params
data frame represents.}
\usage{
transform(params, colspec)
}
\value{
A new data frame consisting of transformed parameter values
}
\description{
This function takes a params data frame consisting of sampled parameter values
and a colspec data frame that specifies what each column in the params
data frame represents.
}
